refactor: Extract association helpers
